pyproject.toml
[project]
name = "openaq-anomaly-prediction"
version = "0.1.0"
description = "A data science project to detect and predict anomalies in the air quality from OpenAQ data."
authors = [
    {name = "Denis CCK", email = "2627434+denis-cck@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12.0, <3.13"

dependencies = [
    "loguru (>=0.7.3,<0.8.0)",
    "numpy (>=2.3.5,<3.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "matplotlib (>=3.10.7,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "openaq (>=0.6.0,<0.7.0)",
    "mlflow (>=3.7.0,<4.0.0)",
    "boto3 (>=1.42.5,<2.0.0)",
    "pyarrow (>=22.0.0,<23.0.0)",
    "duckdb (>=1.4.3,<2.0.0)",
    "statsforecast (>=2.0.3,<3.0.0)",
    "ipykernel (>=7.1.0,<8.0.0)",
    "plotly (>=6.5.0,<7.0.0)",
    "nbformat (>=5.10.4,<6.0.0)",
    "ipywidgets (>=8.1.8,<9.0.0)",
    "google-cloud-bigquery (>=3.39.0,<4.0.0)",
    "dbt-core (>=1.11.2,<2.0.0)",
    "dbt-bigquery (>=1.11.0,<2.0.0)",
]

[dependency-groups]
dev = [
    "ruff (>=0.14.8,<0.15.0)",
    "ipykernel (>=7.1.0,<8.0.0)"
]


# OTHER WAYS ========================================

# PIP
# Dev dependencies for pip install -e ".[dev]"
# [project.optional-dependencies]
# dev = []

# PIP OR UV
# Make the package in src/ available for import with `pip install -e .`
# [build-system]
# requires = ["setuptools>=61.0.0"]
# build-backend = "setuptools.build_meta"
# [tool.setuptools.packages]
# find = {where = ["src"]}